wallet: move crypter to wallet
