+++ src/wallet/crypter.cpp
