+++ src/wallet/wallet.h
@@ -9,11 +9,11 @@
 #include "amount.h"
 #include "primitives/block.h"
 #include "primitives/transaction.h"
-#include "crypter.h"
 #include "key.h"
 #include "keystore.h"
 #include "main.h"
 #include "ui_interface.h"
+#include "wallet/crypter.h"
 #include "wallet/wallet_ismine.h"
 #include "wallet/walletdb.h"
